+++ src/components/Panes/Help/Help.tsx
@@ -1,7 +1,9 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { Pane } from '@common-types';
+import { Pane, Word as WordInterface } from '@common-types';
+
+import { getWordReportForMultipleWords } from '@api/getWordReport';
 
 import usePanes from '@hooks/usePanes';
 import useScrollEffect from '@hooks/useScrollEffect';
@@ -10,47 +12,87 @@ import Word from '@components/Words/Word'
 import Button from '@components/Button/Button';
 import IconGamepad from '@components/Icons/IconGamepad';
 
-import { HELP_WORDS, HELP_WORDS_ALT } from './constants';
+import { HELP_EXAMPLES_BY_LANG } from './constants';
 
 import './Help.scss'
 
 const Help = () => {
+    const [helpGuesses, setHelpGuesses]= useState<WordInterface[]>([]);
     const [isAlt, setIsAlt] = useState(false);
-    const { t } = useTranslation();
+
+    const { t, i18n } = useTranslation();
+
     const { changePane } = usePanes();
 
-    const words = isAlt ? HELP_WORDS_ALT : HELP_WORDS;
     const tEnd = isAlt ? `Alt` : '';
 
+    useEffect(() => {
+        (async () => {
+            const language = i18n.language;
+            if (['pl', 'en'].includes(language)) {
+                const keyToUse = isAlt ? 'second' : 'first';
+                const {
+                    wordToGuess = '',
+                    words = [],
+                } = HELP_EXAMPLES_BY_LANG[language] ? HELP_EXAMPLES_BY_LANG[language][keyToUse] : {};
+
+                if (words) {
+                    const { results } = await getWordReportForMultipleWords(wordToGuess, words, { lang: language, shouldCheckIfExist: false });
+
+                    const guesses = results.map(({ word = '', affixes = [] }) => ({
+                        word,
+                        affixes,
+                    }));
+    
+                    setHelpGuesses(guesses);
+
+                    return;
+                }   
+            }
+            
+            setHelpGuesses([]);
+        })();
+    }, [isAlt, i18n.language]);
+
     useScrollEffect('top', [isAlt]);
+
+    const [
+        wordExampleNotMatch,
+        wordExampleOrder,
+        wordExampleInRow,
+        wordExampleStartAndEnd,
+        wordExampleWin,
+    ] = helpGuesses;
 
     return (
         <div className="help">
             <h2 className="title">{t('help.howToPlayTitle')}</h2>
             <p>{t('help.howToPlayText1')}</p>
             <p>{t('help.howToPlayText2')}</p>
+            {helpGuesses.length > 0 && <>
             <h2 className="title">{t(`help.exampleTitle${tEnd}`)}</h2>
-            <Word guess={words[0]} />
-            <p>{t('help.incorrectLettersTip')}</p>
-            <Word guess={words[1]} />
-            <p dangerouslySetInnerHTML={{ __html: t(`help.correctLettersTip${tEnd}`) }}></p>
-            <Word guess={words[2]} />
-            <p dangerouslySetInnerHTML={{ __html: t(`help.sequenceOfLettersTip${tEnd}`) }}></p>
-            <Word guess={words[3]} />
-            <p dangerouslySetInnerHTML={{ __html: t(`help.firstAndLastLetterTip${tEnd}`) }}></p>
-            <Word guess={words[4]} />
-            <p>{t('help.winingWordMessage')}</p>
-            <p>
-                <Button onClick={() => changePane(Pane.Game)} isLarge>
-                    <IconGamepad />
-                    <span>{t('common.play')}</span>
-                </Button>
-            </p>
-            <p>
-                <Button onClick={() => setIsAlt(value => !value)} isInverted isText hasBorder={false}>
-                    <span>{t(isAlt ? 'help.previousExample' : 'help.altExample')}</span>
-                </Button>
-            </p>
+                <Word guess={wordExampleNotMatch} />
+                <p>{t('help.incorrectLettersTip')}</p>
+                <Word guess={wordExampleOrder} />
+                <p dangerouslySetInnerHTML={{ __html: t(`help.correctLettersTip${tEnd}`) }}></p>
+                <Word guess={wordExampleInRow} />
+                <p dangerouslySetInnerHTML={{ __html: t(`help.sequenceOfLettersTip${tEnd}`) }}></p>
+                <Word guess={wordExampleStartAndEnd} />
+                <p dangerouslySetInnerHTML={{ __html: t(`help.firstAndLastLetterTip${tEnd}`) }}></p>
+                <Word guess={wordExampleWin} />
+                <p>{t('help.winingWordMessage')}</p>
+                <p>
+                    <Button onClick={() => changePane(Pane.Game)} isLarge>
+                        <IconGamepad />
+                        <span>{t('common.play')}</span>
+                    </Button>
+                </p>
+                <p>
+                    <Button onClick={() => setIsAlt(value => !value)} isInverted isText hasBorder={false}>
+                        <span>{t(isAlt ? 'help.previousExample' : 'help.altExample')}</span>
+                    </Button>
+                </p>            
+            </>}
         </div>
     )
 };